Add DraftVocabMixin
