+++ src/speculators/models/eagle3/config.py
@@ -1,7 +1,7 @@
-from typing import Literal
+from typing import Any, Literal
 
-from pydantic import Field
-from transformers import PretrainedConfig
+from pydantic import Field, field_serializer, field_validator
+from transformers import AutoConfig, PretrainedConfig
 from transformers.models.llama.configuration_llama import LlamaConfig
 
 from speculators import SpeculatorModelConfig
@@ -68,6 +68,24 @@ class Eagle3SpeculatorConfig(SpeculatorModelConfig):
         default=False,
         description="Whether embedding layer weights require gradients during training",
     )
+
+    @field_serializer("transformer_layer_config")
+    def serialize_transformer_config(self, value: PretrainedConfig) -> dict:
+        """Serialize transformer config to dict."""
+        return value.to_diff_dict()
+
+    @field_validator("transformer_layer_config", mode="before")
+    @classmethod
+    def validate_transformer_config(cls, value: Any) -> PretrainedConfig:
+        """Validate and convert transformer config."""
+        if isinstance(value, dict):
+            config_class: type[PretrainedConfig] = LlamaConfig
+            if "model_type" in value:
+                config_class = AutoConfig.for_model(
+                    model_type=value["model_type"]
+                ).__class__
+            return config_class(**value)
+        return value
 
     @property
     def target_vocab_size(self) -> int:
